Moved `Sprite` to `MonoGame.Extended.Graphics`
