+++ source/MonoGame.Extended/Sprites/SpriteExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using MonoGame.Extended.Graphics;
 
 namespace MonoGame.Extended.Sprites
 {
