+++ tests/MonoGame.Extended.Entities.Tests/AspectBuilderTests.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using Microsoft.Xna.Framework.Graphics;
-using MonoGame.Extended.Sprites;
+using MonoGame.Extended.Graphics;
 using Xunit;
 
 namespace MonoGame.Extended.Entities.Tests
